Separated the framework strategy content from the framework page
