+++ react-frontend/src/components/DigitalFramework/progress.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Progress = () => {
+  return (
+    <div>
+      <p>progress page filler</p>
+    </div>
+  );
+};
+
+export default Progress;
